% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdapSide.R
\name{AdapSide}
\alias{AdapSide}
\title{An adaptive test leveraging side information}
\usage{
AdapSide(
  Y,
  G,
  cor.est = c("pdsoft", "pdsoft.cv"),
  lam,
  weights,
  pow.param = c(0:10)
)
}
\arguments{
\item{Y}{The outcome variable of interest; a numeric vector of length n.}

\item{G}{a numeric n by J genotype matrix.}

\item{cor.est}{The method for estimating a positive definite correlation matrix.}

\item{lam}{The tuning parameter for estimating correlation matrix; Could be either a scalar or
a p by p symmetric matrix with an irrelevant diagonal while \code{pdsoft} is selected;
Should be a vector when \code{pdsoft.cv} is selected.
See \code{pdsoft} and \code{pdsoft.cv} in \code{PDSCE} for details.}

\item{weights}{a numeric vector of length J, contains the weights from side information.}

\item{pow.param}{an integer numeric vector specifying the \eqn{\gamma} values used in the test statistics.}
}
\value{
A vector of p values.
}
\description{
The function returns the p values of an adaptive test by leveraging available side information.
}
\examples{
 J = 100
 n = 50
 beta=c(rep(2,3),rep(0,J-3))
 data=example_data_generate(n, J, beta)
 Y=data$Y
 G=data$G
 lam=0.1
 weights=c(J:1)
 pow.param=c(1:5)
 AdapSide(Y, G, cor.est="pdsoft", lam, weights, pow.param)


}
\references{
Zhao and Sun (2020). A stable and adaptive polygenic signal detection method based on repeated sample splitting.
   \emph{arXiv:2008.02442}.
}
\keyword{Adaptive}
\keyword{Side}
\keyword{information,}
\keyword{test}
